requirements.txt
openai
youtube-transcript-api
python-dotenv